requirements.txt
flask
gunicorn
psycopg2-binary
python-dotenv
azure-storage-blob